requirements.txt
openai==2.14.0
openpyxl==3.1.5
python-dotenv==1.2.1
PyPDF2==3.0.1
pdf2image==1.17.0
natsort

